ceph/005: skip test if using "test_dummy_encryption"
When using the "test_dummy_encryption" mount option, new file and directory
names will be encrypted.  This means that if using as a mount base directory
a newly created directory, we would have to use the encrypted directory name
instead.  For the moment, ceph doesn't provide a way to get this encrypted
file name, thus for now simply skip this test.

Signed-off-by: Luís Henriques <lhenriques@suse.de>
Reviewed-by: Xiubo Li <xiubli@redhat.com>
Signed-off-by: Zorro Lang <zlang@kernel.org>