fstests: btrfs/219: remove it from auto group
The test case is to make sure we can mount a fs with older generation
(but with the same fsid/dev uuid).

Normally we will reject such case as btrfs is maintaining an internal
devices list (for multi-device support), and if we find a device
suddenly got an older generation, we will directly reject it.

Although for single device btrfs, we may add an exception for it,
the corresponding kernel patch is never merged.

So for now, just remove the test case from auto group.

Signed-off-by: Qu Wenruo <wqu@suse.com>
Reviewed-by: Johannes Thumshirn <johannes.thumshirn@wdc.com>
Signed-off-by: Zorro Lang <zlang@kernel.org>